+++ github/git.go
@@ -6,7 +6,7 @@ import (
 	"gopkg.in/src-d/go-git.v4/storage/memory"
 	"io/ioutil"
 
-	"gitrob/common"
+	"wraith/common"
 
 	"gopkg.in/src-d/go-git.v4"
 	"gopkg.in/src-d/go-git.v4/plumbing"
@@ -27,7 +27,7 @@ func CloneRepository(cloneConfig *common.CloneConfiguration) (*git.Repository, s
 	var err error
 	var dir string
 	if !*cloneConfig.InMemClone {
-		dir, err := ioutil.TempDir("", "gitrob")
+		dir, err := ioutil.TempDir("", "wraith")
 		if err != nil {
 			return nil, "", err
 		}