+++ CHANGELOG.md
@@ -7,6 +7,16 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+- change internal name from gitrob to wraith
+
+### Fixed
+- bug in scanLocalGitRepo configuration flags
+
+## [0.0.3] - 2020-08-06
+### Changed
+- massive rewrite and uplift of code from codeemitter and mattyjones
+
 ## [0.0.2] - 2020-05-27
 ### Fixed
 - fix issue where if a token was not found or invalid it would panic
@@ -16,6 +26,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - initial release for the new project
 
-[Unreleased]: https://github.com/mattyjones/gitrob/compare/0.0.2...HEAD
+[Unreleased]: https://github.com/mattyjones/wraith/compare/0.0.3...HEAD
+[0.0.2]: https://github.com/mattyjones/wraith/releases/tag/0.0.3
 [0.0.2]: https://github.com/mattyjones/gitrob/releases/tag/0.0.2
 [0.0.1]: https://github.com/mattyjones/gitrob/releases/tag/0.0.1